tests/submission/Ms2ger/dynamic-markup-insertion/document.open-02.html

<!DOCTYPE html>
<title>document.open with three arguments</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#dom-document-open">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
function open() {
  assert_unreached("The call should be redirected to the real window.open")
}
test(function() {
  var w = document.open("/resources/testharness.js", "", "");
  assert_true(w instanceof w.Window, "Expected a window");
}, "document.open should redirect to window.open when called with three arguments");
</script>
